129 - Corrigindo Bug da home duplicada, orientação do App e Delete de Imagem IOS

+++ lib/models/product.dart
@@ -7,19 +7,19 @@ import 'package:uuid/uuid.dart';
 import 'dart:io';
 
 class Product extends ChangeNotifier {
-  Product({this.id, this.name, this.description, this.images, this.sizes}) {
+
+  Product({this.id, this.name, this.description, this.images, this.sizes}){
     images = images ?? [];
     sizes = sizes ?? [];
   }
 
-  Product.fromDocument(DocumentSnapshot document) {
+  Product.fromDocument(DocumentSnapshot document){
     id = document.documentID;
     name = document['name'] as String;
     description = document['description'] as String;
     images = List<String>.from(document.data['images'] as List<dynamic>);
-    sizes = (document.data['sizes'] as List<dynamic> ?? [])
-        .map((s) => ItemSize.fromMap(s as Map<String, dynamic>))
-        .toList();
+    sizes = (document.data['sizes'] as List<dynamic> ?? []).map(
+            (s) => ItemSize.fromMap(s as Map<String, dynamic>)).toList();
   }
 
   final Firestore firestore = Firestore.instance;
@@ -38,21 +38,21 @@ class Product extends ChangeNotifier {
 
   bool _loading = false;
   bool get loading => _loading;
-  set loading(bool value) {
+  set loading(bool value){
     _loading = value;
     notifyListeners();
   }
 
   ItemSize _selectedSize;
   ItemSize get selectedSize => _selectedSize;
-  set selectedSize(ItemSize value) {
+  set selectedSize(ItemSize value){
     _selectedSize = value;
     notifyListeners();
   }
 
   int get totalStock {
     int stock = 0;
-    for (final size in sizes) {
+    for(final size in sizes){
       stock += size.stock;
     }
     return stock;
@@ -64,86 +64,73 @@ class Product extends ChangeNotifier {
 
   num get basePrice {
     num lowest = double.infinity;
-    for (final size in sizes) {
-      if (size.price < lowest && size.hasStock) lowest = size.price;
+    for(final size in sizes){
+      if(size.price < lowest && size.hasStock)
+        lowest = size.price;
     }
     return lowest;
   }
 
-  ItemSize findSize(String name) {
+  ItemSize findSize(String name){
     try {
       return sizes.firstWhere((s) => s.name == name);
-    } catch (e) {
+    } catch (e){
       return null;
     }
   }
 
-  List<Map<String, dynamic>> exportSizeList() {
+  List<Map<String, dynamic>> exportSizeList(){
     return sizes.map((size) => size.toMap()).toList();
   }
 
   Future<void> save() async {
     loading = true;
+
     final Map<String, dynamic> data = {
       'name': name,
       'description': description,
       'sizes': exportSizeList(),
     };
 
-    if (id == null) {
+    if(id == null){
       final doc = await firestore.collection('products').add(data);
       id = doc.documentID;
     } else {
       await firestoreRef.updateData(data);
     }
 
-    // IMAGES [URL1, URL2, URL3]
-    // NEWIMAGES [URL2, URL3, FILE1, FILE2]
-    // UPDATED [URL2, URL3, FURL1, FURL2]
-
-    // MANDA FILE1 PRO STORAGE -> FURL1
-    // MANDA FILE2 PRO STORAGE -> FURL2
-    // EXCLUI IMAGEM URL1 DO STORAGE
-
-    // IMAGES [URL1, URL2, URL3]
-    // NEWIMAGES [URL3, FILE1]
-    // UPDATE [URL3, FURL1]
-
-    // MANDA FILE1 PRO STORAGE -> FURL1
-    // EXLUIR URL1 DO STORAGE
-    // EXLUIR URL2 DO STORAGE
-
     final List<String> updateImages = [];
 
-    for (final newImage in newImages) {
-      if (images.contains(newImage)) {
+    for(final newImage in newImages){
+      if(images.contains(newImage)){
         updateImages.add(newImage as String);
       } else {
-        final StorageUploadTask task =
-            storageRef.child(Uuid().v1()).putFile(newImage as File);
+        final StorageUploadTask task = storageRef.child(Uuid().v1()).putFile(newImage as File);
         final StorageTaskSnapshot snapshot = await task.onComplete;
         final String url = await snapshot.ref.getDownloadURL() as String;
         updateImages.add(url);
       }
     }
 
-    for (final image in images) {
-      if (!newImages.contains(image)) {
+    for(final image in images){
+      if(!newImages.contains(image) && image.contains('firebase')){
         try {
           final ref = await storage.getReferenceFromUrl(image);
           await ref.delete();
-        } catch (e) {
+        } catch (e){
           debugPrint('Falha ao deletar $image');
         }
       }
     }
 
     await firestoreRef.updateData({'images': updateImages});
+
     images = updateImages;
+
     loading = false;
   }
 
-  Product clone() {
+  Product clone(){
     return Product(
       id: id,
       name: name,
